修改接收服务器逻辑 由session转为一个一个ReceivedManager类专门处理

+++ ETImClient/ETImClient/Module/Account/ViewController/OnlineViewController.mm
@@ -52,7 +52,7 @@ using namespace etim::pub;
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleSubtitle reuseIdentifier:identifier];
     }
     
-    cell.textLabel.text = [NSString stringWithFormat:@"%d", indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"%ld", (long)indexPath.row];
     
     return cell;
 }
